Move classes out of top-level package
