+++ app/src/main/java/org/mozilla/fenix/ext/Context.kt
@@ -25,7 +25,7 @@ import mozilla.components.support.base.log.Log
 import mozilla.components.support.base.log.Log.Priority.WARN
 import org.mozilla.fenix.FenixApplication
 import org.mozilla.fenix.R
-import org.mozilla.fenix.ThemeManager
+import org.mozilla.fenix.theme.ThemeManager
 import org.mozilla.fenix.components.Components
 import org.mozilla.fenix.components.metrics.MetricController
