+++ app/src/main/java/org/mozilla/fenix/library/history/HistoryItemMenu.kt
@@ -8,7 +8,7 @@ import android.content.Context
 import mozilla.components.browser.menu.BrowserMenuBuilder
 import mozilla.components.browser.menu.item.SimpleBrowserMenuItem
 import org.mozilla.fenix.R
-import org.mozilla.fenix.ThemeManager
+import org.mozilla.fenix.theme.ThemeManager
 import org.mozilla.fenix.library.LibraryItemMenu
 
 class HistoryItemMenu(
